ai-service/requirements.txt
livekit-agents
livekit-api
opencv-python
numpy
python-dotenv
aio-pika
